move etcd-backup to the CPO
