+++ Containerfile.operator
@@ -21,7 +21,6 @@ COPY --from=builder \
   /opt/app-root/src/bin/hypershift-operator \
   /opt/app-root/src/bin/control-plane-operator \
   /opt/app-root/src/bin/control-plane-pki-operator \
-  /opt/app-root/src/bin/etcd-backup \
   /usr/bin/
 
 RUN cd /usr/bin && \
