+++ etcd-backup/OWNERS
@@ -1,2 +1,2 @@
 labels:
-- area/hypershift-operator
+- area/control-plane-operator
